moved button and checkboxes to the GameParameters
